Image rotation
Adds a button to the image viewer control frame to rotate the image by 90º counter-clockwise. If an image has been rotated, the rotation will be stored in the JSON file to persist across restarts. Images that haven't been rotated aren't stored in this data and assumed to be rotated 0º.

Fixes #953