Add regex timeout
Some regexes can take effectively for ever, so add a
timeout, settable via an advanced Pref.

If regex times out during search, disable highlighting
of search string temporarily, or it will keep trying to
repeat the failing search.

Error message may appear twice, once for the search
and once for the highlighting. I think this is an
acceptable price to pay - alternatives such as setting
a flag telling searches to be ignored would run a
risk of breaking things. This is to fix a very rare
occurrence in a way that doesn't risk breaking
other user searches or intra-code searching.