Improve a few scanno regexes
Five scannos with hints "Find a string that starts with
xx and change it to yy" are better with a lookahead to
check there is a character after it. This stops them
matching "tb" (thought break) or "hr" (horizontal rule).

Fixes #606

Testing notes. Add words that begin with hl, hr, rn,
tb, and tli, as well as those strings as standalone words.
The standalone words should no longer be reported
by Stealth Scannos, regex.json.