Investigate tk9 Mac scroll
Testing notes:

In Tk8, you should see a debug print "Failed to bind
TouchpadScroll". It would be interesting to know if
Tk9 gives the same message.

In Tk8 and Tk9, attempt to scroll the minimum amount
in the image viewer, using the mousewheel.
You should see a debug print "delta: <nn>" in the shell
window. When I scroll the mouse wheel one notch, I
get a delta of +/- 120.
In Tk8, I think you should get a delta of +/- 1.
It would be interesting to know what Tk9 gives.

It's possible that using touchpad scrolling will give
lots of messages "Touchpad: <nn>". Again, it would
be interesting to know what happens there.