Add Commonly Used Chars to Unicode menu (#1003)
It's really the Unicode Block dialog, but with the
dropdown block selection set to Commonly Used
Chars.

If you change blocks, it will be stored in the pref,
and when you next do Unicode Block, it will show
that block, but if you do Common Chars, it will
force to that. If you change the block to be
Common Chars, that will be stored in the pref,
and that's what you'll get next time you do pop
Unicode Block.

Also added Shift left/right click to status bar, where
the character name is shown - instructions in tooltip.