Improvements to Stealth Scannos (#1001)
* Improve a few scanno regexes

Five scannos with hints "Find a string that starts with
xx and change it to yy" are better with a lookahead to
check there is a character after it. This stops them
matching "tb" (thought break) or "hr" (horizontal rule).

Fixes #606

Testing notes. Add words that begin with hl, hr, rn,
tb, and tli, as well as those strings as standalone words.
The standalone words should no longer be reported
by Stealth Scannos, regex.json.

* Improve scannos tooltip

It hadn't been updated to describe using the Shift
key to "fix 'em all"

* Add orphan commands to scannos dialog

* Further scanno regex improvements

1. "guess" exception extended to  include "(mar)quess"
2. Punct at start of line not triggered by ellipsis
3. "ii" only suspicious if preceded by a, e, o, u.
4. Ellipsis followed by lower case not reported

* Improve "tb" check number 41