Fix tabbing through statusbar/image viewer (#1027)
Fixes #1026

The order is main text, peer window (if visible), status bar, image viewer (if visible and docked), back to main
text. If image viewer is floating, it is not included in
the tab loop from the main window, but has its own
tab loop. You need to Ctrl/Cmd+tab or click in the
image viewer to get focus in the window, then tabbing
will just cycle through the image viewer buttons. In
short, the floating image viewer should behave like a
dialog as far as tabbing is concerned.

Testing notes - there are 6 cases to test:
1. No split text or image viewer
2. Just split text
3. Just docked image viewer
4. Just undocked image viewer
5. Split text and docked image viewer
6 Split text and undocked image viewer

Transition between regions is the most likely place
for things to go wrong, e.g. at beginning/end of
status bar, or beginning/end of image viewer.