Improve keyboard support for "Apply"
Bind "Shift+Return" to "Apply" button in an
Ok, Apply, Cancel dialog, in the same way as "Return"
activates the "OK" button, and "Escape" activates
the "Cancel" button.

Also, add "Apply" in Surround Selection With" as a
shadow command, so it can be activated when the
dialog doesn't have focus (because user has just
selected the text they want to surround).
This latter bit fixes #1007