Add shadow commands for image zoom (#1035)
* Add shadow commands for image zoom

See all the "Image" shadows by typing "Image" into
Command Palette.

This adds zoom in/out (Ctrl/Cmd-plus/minus) and
fit to height (but not autofit) (Ctrl/Cmd-zero).
Those were the default shortcuts we already had, but
this adds them as shadow commands, so the user can
reassign them.

Test in docked and undocked state, and when image
is not shown.

* Fix "minus", and add Ctrl-equals shortcut