Make Ctrl-Z refresh current page separator
When Page Sep dialog is being used, the Undo/Redo
buttons in the dialog (and dialog shortcuts) not only
performed undo/redo, but also refreshed the display
to show the current page separator. However, using
Undo/Redo from the main Edit menu, or Ctrl-Z, etc.
when focused in maintext did not do the refresh.

This commit fixes that, so that however the Undo is
activated, the Refresh will happen. Once the dialog is
closed, the additional refresh will no longer occur.

Fixes #938