Improve WF navigation (#1043)
1. If no more occurrences of word, reset current insert
position to start/end of file (depending on direction)
similar to GG1, as well as ringing the bell.
2. Allow Shift-click on word to search backward from
current location - finds last occurrence on first click
3. Fix bug where Home/End/Cmd+Up/Down keys
didn't show the first occurrence of a word if that word
was already highlighted, for example, if pressing Home
twice, the second press went to the second occurrence
of the word, instead of always the first one.