Add Highlight All to S/R dialog (#1040)
This is a suggested solution for the "Highlight Char,
String & Regex" GG1 feature.

Use the S/R dialog to set Regex, wholeword, case
sensitivity, and optionally "In Selection". Then use
Highlight All to highlight all matches. The settings, such
as wholeword are "frozen" for the highlighting, i.e. if
the user then changes the settings or the search string,
it doesn't change the highlighting.

Like GG1, the color is the same as the quotes
highlighting. Remove highlighting by using the
Search->Remove Highlights button.

New button has a shadow command, like the other
S/R buttons.