Add "Two Words" as option in WF Hyphens
In master, "Two Words" is always on, so "flash light"
will be queried as a possible suspect with "flash-light".

This commit makes that optional, with a checkbox
next to the Hyphens radiobutton in WF. This is similar
to GG1, except there it is in the Prefs menu, not the
WF dialog.

(Note "flashlight" as one word is always queried, and
is unchanged by this commit.

Fixes #925