Catch repeated word at line start with spaces
Repeated words are reported if the first is at the end
of a line, and the second is at the start of the next line.
However, the regex didn't allow the second word to
have spaces before it, which it might for an indented
piece of text, e.g. blockquote.