Fix repeated words bug
A line with two distinct repeats, e.g.
`dog dog cat dog dog`
was reported as having 3 repeats, due to a bug in
the way the first word in each double match was
cleared to avoid re-finding it.
It was clearing the first word on the line, not the
first in the pair, e.g. after first match
`     dog cat dog dog`
After second match
`        cat dog dog`
instead of
`    dog cat     dog`