PPtxt improvements (#1054)
* Suppress multiple identical repeated word errors

In PPtxt, if a line contains a word several times, e.g. in
a table heading, we only want one report per line

* Catch repeated word at line start with spaces

Repeated words are reported if the first is at the end
of a line, and the second is at the start of the next line.
However, the regex didn't allow the second word to
have spaces before it, which it might for an indented
piece of text, e.g. blockquote.

* Fix mypy complaints

Update versions to match Github actions

* Fix repeated words bug

A line with two distinct repeats, e.g.
`dog dog cat dog dog`
was reported as having 3 repeats, due to a bug in
the way the first word in each double match was
cleared to avoid re-finding it.
It was clearing the first word on the line, not the
first in the pair, e.g. after first match
`     dog cat dog dog`
After second match
`        cat dog dog`
instead of
`    dog cat     dog`

* Comment update

* Don't report spaced close curly quotes if dittos

If close curly quote has 2 spaces on each side (start/end
of line also counts as 2 spaces) it's ditto mark, not a
badly spaced quote.

* Treat '----' similarly to '--'

These are used in text files to represent emdashes.

If '--` occurs more than 5 times, a message was output
that future occurrences won't be reported. The same
has now been added for `----`

* Improve reporting of standalone 0 & 1

E.g. report ` .0` or `0.234`

Also, consolidate repeated "special situations checks",
for example don't report each spaced period individually
in `. . . . . . . . . . .`