Make curly quote exception optional
Previously if a closing curly quote was missing at the
end of a paragraph, no error was output if there was
an open quote at the start of the next paragraph. This
was to avoid many false positives from this common
occurrence. This feature is in PPWB's ppsmq, and in
GG1.

However, it's not possible to know that the missing
quote was missing at the END of the paragraph, e.g.
`"Hello, he said.` So, this exception could cause
true errors to be omitted.

This commit makes the exception optional via a
Preference controlled by a checkbox in the Curly
Quote Check dialog.

Fixes #1052