Let user decide whether to build epub2 file (#1060)
When running ebookmaker locally or via the API, things
fail badly if it is not told to build at least one format.
Previously, GG forced the user to build EPUB2, but with
this commit, they can choose which formats to build
and receive a polite error message if they don't choose
any formats.