Make Prev/Next FN work if anchor message selected (#1086)
Previously if the selected line in the FN dialog was an
anchor line rather than a footnote line, you couldn't do
Prev/Next FN. Selecting an anchor line is now the same
as selecting the footnote it belongs to.

Fixes #1068 - see issue for example file and description
of how bug was noticed by user.