Fix Illo Fixup move bug
If two illos are adjacent (one blank line between) and the
upper one has blank lines, i.e. multi paragraph, and the
user tries to move the lower one upwards, it is supposed
to do nothing because you can't move illos past one another,
but if finds the blank line in the upper illo and tries to move
it in there, potentially losing lines in the process.

This fix makes it detect the end of illo markup, then it
tracks back to find the start of illo markup, before beginning
the search process. If indeed it was an illo, it will do nothing,
but if it was a sidenote, it can move the illo above the
sidenote.

Hopefully fixes #1143