Fix Illo Fixup move bug (#1148)
* Fix Illo Fixup move bug

If two illos are adjacent (one blank line between) and the
upper one has blank lines, i.e. multi paragraph, and the
user tries to move the lower one upwards, it is supposed
to do nothing because you can't move illos past one another,
but if finds the blank line in the upper illo and tries to move
it in there, potentially losing lines in the process.

This fix makes it detect the end of illo markup, then it
tracks back to find the start of illo markup, before beginning
the search process. If indeed it was an illo, it will do nothing,
but if it was a sidenote, it can move the illo above the
sidenote.

Hopefully fixes #1143

* Small bug in previous fix

Didn't detect single line markups above the footnote
being moved up, e.g. `[Blank Page]`

* Fix illo moving bug

1. Avoid double interpretation of start/end points from
marks to indexes, which could cause exception
2. Minor adjustment also needed to ensure asterisk
before markup dealt with correctly

* Fix bug in earlier commit

When a line ended in `[3]` for example, the code mistook
it for illo markup.

Also, added bell when illo/sidenote can't be moved.

* Minor bug fix