Don't break leading `/#` markup in HTML gen (#1158)
* Don't break leading `/#` markup in HTML gen

If file begins with `/#` then HTML generation fails because
the `/#` ends up on a line with other markup. Then later,
this open markup is not recognized, and a fatal error occurs
when the matching close markup is found.

This is a quick fix to let HTML generation continue, rather
than causing a fatal error.

#1157 addresses the issue more fully.

* Note in changelog