Add 'Replace Match' functionality to search menu (#1159)
Replaces the latest found match with the latest replacement
string. If the S/R dialog is currently visible, then just like
Find Next/Previous, it uses the values from the fields in
that dialog, even if they haven't been used for search &
replace yet.

Errors, such as trying to replace before doing a search
should ring the bell. If the regex is invalid, that will
pop an error dialog if the S/R dialog is not available to
display the regex error message.