Stop Proofer Comment deletion removing edits (#1163)
If the PPer corrected the word before a proofer comment
by typing letters immediately before (or after) the comment,
then used the Delete function of the Proofer Comment
dialog, their edit could be lost or corrupted.

Avoid by using RIGHT gravity for the mark before the
comment and LEFT gravity afterward, which it the
opposite to default checker behavior.