Add variable multi-replace to SR dialog (#1169)
User can now change the number of replace fields shown
in the SR dialog between 2 and 10, default is 3.

Dialog should resize correctly, and setting is stored in
prefs, so should be persistent.

Also fixed minor bug relating to multi-replace. In master,
if a replacement string is used in replace field 1, for example,
it doesn't appear in the drop down for replace field 2
until the dialog is closed and re-opened. With this commit,
strings are added to drop-down history for all fields as
soon as they are used in any field.

Fixes #792