Warn if nested bracket confuses Footnote Fixup
A nested square bracket within a continued footnote can
cause Footnote Fixup to combine two footnotes into one.

This is a simple fix to report this situation by detecting that
and unexpected `[Footnote` appears within the text of a
footnote.

A more robust search method or fix may transpire when
the work is done for #515, but this will at least alert PPers.
This is quite a rare event, but if it happens, the PPer can
either copy/paste the continuation text into the correct place, or close the nested square bracket temporarily until footnote
fixup is complete.

Fixes #1155