Improve WF whole-word searching
If `Gaume` and `Gaume's` both appear in a text, they are
reported in WF as separate words. However, when `Gaume`
is clicked, it also matched `Gaume's` in the text (due to
regex wordchar/non-wordchar inconsistencices).

This commit improves on that behavior. Now, clicking
`Gaume` will find `Gaume` and `Gaume'` (closing straight
quote) and `Gaume’` (closing curly quote), but will
not match `Gaume's` or `Gaume’s` (possessive apostrophes).
Apostrophes at the start of the word are treated similarly,
i.e.  it will no longer match `O'Gaume` or `O’Gaume`.

Fixes #1185