Add less strict curly quote conversion (#1213)
* Add less strict curly quote conversion

Add a preference to the Settings, Advanced Tab, for
single quote conversion to be "strict" (i.e like `master`)
or not.

If not, and a single quote is found within double quotes
in the same paragraph, and the single quote as at the
start of a word, it is assumed to be an apostrophe and
converted to a curly one.

Fixes #671

* Allow apostrophe'd dialect to be preceded by DQ

Also, allow `>` (DP markup)