Add multi-line chapter heading option to HTML gen (#1212)
* Add multi-line chapter heading option to HTML gen

If on (default) HTML generation will behave like `master`,
including all lines until a double blank line in the h2 heading.
If off, just the first line after the 4 blank lines will be
taken as the h2 heading.

This should be useful for books where there is a blockquote
after the chapter that should not be marked up as h2, or for
some periodicals. See #1109 for discussion.

Fixes #1109

* Make blockquote signify end of chapter heading

BQ would have ended chapter heading if single-line
headings, but not multi-line. This commit causes the
open BQ markup to signify end of heading, then code
loops round again to actually process the BQ markup.