Extend bookmarks to also save current selection (#1228)
When user sets bookmark, if there is a selection, also mark
and save that.
When user restores such a bookmark, the selection will
also be restored.

Also, highlighting of bookmarks (brief flash of green) was
broken in master, so fixed as well.

Also, ensure focus is in main text after using the "restore
selection" button in the toolbar