Make caption para alignment inherit, not justify
By default paragraphs are set to be justified in our books by
the CSS header.

This commit makes the justification follow whatever is set for
the parent, the figcaption, defaults to center on my system.

If the PPer wants some/all paragraphs to have different
alignment, they will need to either use CSS for all such
paragraphs, or for specific captions.