Fix reindex bug if footnote has 1st line footnote (#1242)
If a footnote has a footnote, with the anchor on the first
line of the original footnote, then re-indexing causes
marks to be moved, and the whole first line of the footnote
is replaced with an anchor.

Fixed by only replacing the footnote label when the
reindexing is done.

Fixes #1238