Add option to use `<p>` markup in captions (#1237)
* Add option to use `<p>` markup in captions

In the Image/Auto-Illus dialog, provide a checkbutton to
allow PPer to select whether they want `<p>` markup
for captions.

Also add CSS to default header to remove the top/bottom
margin for `<p>` inside captions, or they end up too spaced
out. The PPer can override this either in their own default
header, or for a specific book.

* Review fixes

1. Avoid including leading spaces in `<p>` version of caption,
even if text caption was indented.
2. Add 0.2em bottom margin to caption paragraphs

* Make caption para alignment inherit, not justify

By default paragraphs are set to be justified in our books by
the CSS header.

This commit makes the justification follow whatever is set for
the parent, the figcaption, defaults to center on my system.

If the PPer wants some/all paragraphs to have different
alignment, they will need to either use CSS for all such
paragraphs, or for specific captions.