Improve Add to Project Dict in Spell Check (#1250)
The following now all do the same, which is to add to the
project dict and remove all occurrences from list:
1. Click Add to Project Dict
2. Cmd/Ctrl-P (GG1 shortcut)
3. Cmd/Ctrl-left click
4. Cmd/Ctrl-right click

Button shortcut is unchanged "Cmd/Ctrl-P or Ctrl-click
message", which is now both correct and useful.

The following now all do the same, which is to add to the
global dict and remove all occurrences from list:
1. Click Add to Global Dict
2. Cmd/Ctrl-A (GG1 shortcut)
3. Shift-Cmd/Ctrl-left click
4. Shift-Cmd/Ctrl-right click

Button shortcut is now "Cmd/Ctrl-A or Shift Ctrl-click
message".