Stop checker exception if closed while running (#1268)
Specific example found was with the illo fixup, but I think
it could happen with other checkers. I don't think it's
related to undo/redo.

To reproduce, duplicate enough copies of the text of a file
inside the file, so that illo fixup takes a few seconds to run.
While it is running or re-running, close the checker dialog.

Fixes #1255