Make PPtxt less verbose (#1274)
* Add checkboxes to run groups of tests in PPtxt

* Remove trailing spaces check

* Relax emdash-punctuation exclusion

Don't insist on it being at end of line

* Improve scannos checks

1. Only one report per line for any scanno, e.g. "arid arid"
2. Max 5 reports for each scanno

* Report weird characters less often

1. Only report weirdo once per line, even if several
2. Don't report Greek if more than 5 different Greek chars

* Only report `comma after "the"` 5 times

Often many of these in an index.

Tests are split into comma after "the" and other punctuation
after "the". Only the former has the limit of 5 reports.

* Update test results

* Reduce false positives from period-comma

Split double punctuation checks into period-comma (which
is often valid) and others (which often aren't).

Only report period-comma 5 times.

* Add simple longest/shortest line check like GG1's

This does not do a full analysis like bookloupe, searching for
short lines that are not at the end of a paragraph, but just
a simple report of the max and min line lengths in the file.

It does not seem valuable to further duplicate bookloupe's
behavior in PPtxt

* Reduce reports of emdashes

* Add fractions, super- & subscripts to ok chars

* Update "character checks" heading