Fix removal of blank line for mid-para illo/sn
If there was an asterisk, it broke the code that checked if a
superfluous blank line should be removed once the mid-para
illo/sn was moved.

Also, if a mid-para illo/sn is moved, it is no longer mid-para
but we want it to still be a "suspect" or it vanished from the
Suspects Only list, so mark it as "PREV MIDPARA".

Fixes #1273