Improve Illo/SN movement over multi-blanklines (#1276)
* Improve Illo/SN movement over multi-blanklines

Previously, it skipped over all blank lines in a block when
moving illo/sn up. It is less likely that this is the desired
position when moving up, i.e. to want the sidenote/illo
to have multiple blank lines below it.

So, this commit defaults to just moving to the first blank
line upwards. If the user wants to move further they can
click "Up" again. Requiring an extra click in the less frequent
circumstance seems preferable to not being able to move
the illo/sn where you want to.

Although this means several clicks would be necessary to move an illos/sn past a chapter heading, in practice, it is
uncommon to want to move them to a different chapter.

Fixes #1273

* Fix removal of blank line for mid-para illo/sn

If there was an asterisk, it broke the code that checked if a
superfluous blank line should be removed once the mid-para
illo/sn was moved.

Also, if a mid-para illo/sn is moved, it is no longer mid-para
but we want it to still be a "suspect" or it vanished from the
Suspects Only list, so mark it as "PREV MIDPARA".

Fixes #1273