Allow user to drag selected text to another place
Initial commit to gather feedback...

1. Only drags the first bit of a column selection - one
possibility instead would be to disable dragging if there's
a column selection
2. If you drag below the bottom of the screen, it still works,
but appears to select all the intervening text until you
release the mouse, when the dragged text should appear
in the correct place. Again, may be possible to fix.
3. Dropping should drop immediately before the character
the mouse is released over, unless at the very right of that
character, in which case it should go to the right.

Testing:
Drag small bits of text left and right, as well as larger chunks
up or down.