Add sorting to the Compose Help dialog
1. Display an arrow in the header of the dialog to show how
the list is sorted.
2. Click in the header to sort by that column
3. Click again to reverse sort
4. Sort method is stored in prefs file, so persistent across
runs of the program

Fixes #1021

I don't think a help dialog warrants a search field and button
as I suggested in point 2 of the issue. It's very quick to find
any character if you sort the list by character name. The GG1
equivalent is just a list. In GG2, we now have sorting, and you
can click on items in the list to insert the character, but its
main purpose is to help if you can't remember the compose
sequence for a character you want.