Strip HTML tags from Custom menu $t token
Modify token "$t" available to Custom menu items. This token will
contain the selected text, as before, but now with HTML tags stripped.
The tags aren't validated as valid HTML; anything between `< >` is
stripped out.

Contiguous whitespace is condensed to a single space, as before.

Rationale for this change: when submitting queries to search engines,
Ngrams, Google Translate, etc, the presence of the tags (or more likely,
the text within them) can change the result, or is just invalid input
(as with Ngrams).