Allow user to drag selected text to another place (#1315)
* Allow user to drag selected text to another place

Initial commit to gather feedback...

1. Only drags the first bit of a column selection - one
possibility instead would be to disable dragging if there's
a column selection
2. If you drag below the bottom of the screen, it still works,
but appears to select all the intervening text until you
release the mouse, when the dragged text should appear
in the correct place. Again, may be possible to fix.
3. Dropping should drop immediately before the character
the mouse is released over, unless at the very right of that
character, in which case it should go to the right.

Testing:
Drag small bits of text left and right, as well as larger chunks
up or down.

* Improve dragging a selection

1. Make it work in peer window as well
2. Show insert point with insert cursor during dragging
3. Change mouse cursor shape while dragging
4. Drag (potentially truncated) preview of text with mouse
5. Allow Cmd/Ctrl dragging (copy not move) text to location
immediately before/after selection, e.g. `abc` --> `abcabc`
6. Disable dragging if selection is a column selection

* Add "copy" indicator

When Ctrl/Cmd is held, dragging copies rather than moves.
Show a "+" label next to the dragged text to indicate "copy".

* Further attempts to fix

1. Allow any modifier key during dragging to mean "copy".
On Windows, Ctrl or Shift work. Need to check which Mac
keys work, so it can be documented.
2. When attempt to drag&drop outside window, interceot
<Leave> event and stop other behavior. If mouse is brought
back inside, user can continue to drag & the text and drop
it.
3. Fix made to cancelling drag&drop using the Escape key -
now restores cursor shape correctly.

* Fix cursors