Fix silent crash using Recent Files menu (#1321)
* Fix silent crash using Recent Files menu

When the file name changes, previously the whole menu
system was cleared and re-written, just because the Recent
Documents menu needed rewriting. Either the repeated
clearing and recreating (internal structures or Tk menus) or
occasional timing clashes of using the menu and recreating
it at the same time caused GG to silently exit.

Fixed by rewriting the labels and commands in the Recent
Docs menu and not deleting/recreating everything.

Fixes #1319

* Black fix