Implement "replace identicals" feature (#1316)
When using the Replace All buttons in the S/R dialog, if the
button is Shift-clicked, it only replaces the strings that
exactly match the current match. It then moves on to the
next match.

"Exactly" means not regex (regardless of dialog setting)
because it's an exact match; case sensitive (regardless of
dialog setting) - could be argued it should match dialog;
whole word matches dialog setting (or chaos would ensue)

Fixes #609