Add "type-to-search" to Unicode Block dialog (#1326)
1. Ensure focus is in the dropdown when dialog is created
2. If user types, store the characters they type
3. If they pause between characters, reset and start again
4. If they type a string that matches the start of a word in
the Unicode block name, jump to it. For example, typing
`math` can match `Mathematical Operators` or
`Miscellaneous Mathematical Symbols`
5. If they type a string that is the start of a 4-digit hex
unicode codepoint, go to the block that contains that
codepoint, if it is in the list. For example, if they know that
open single quote is `U+2018`, they can type `201` or `2018`
and it will find the `General Punctuation` block which has
range 2000-206f.

Fixes #944