Add "Fix/Hide/Fix&Hide All" buttons to Scannos
This is mainly for consistency. Other tools work via the
Hide/Fix All buttons. Although Fix All is equivalent to Replace
All in Stealth Scannos, it caused confusion that there was
no Fix All button. We are retaining the Replace/Replace All
buttons so that they are near the replacement fields and
are more like GG1, particularly where Scannos is (ab)used
to operate like a S/R dialog with custom S/R strings.

Fixes #1175