Add advanced Home key behavior
If Home (or Cmd-Left) is pressed, go to start of line. If already
at start of line, then go to first non-space on current line.
If Shift is pressed then make selection or adjust start/end
of existing selection to "home" position depending on
where current insert position is.

Fixes #1160