Use `blockquote` instead of `div class="blockquot"`
For reasons linked from the issue, it's better to use the
HTML blockquote tag. If a PPer really wants to use divs
instead, a simple search/replace will suffice. This would
not be true in reverse if the PPer wanted blockquotes
and divs had been used, because the correct matching
`/div` tags are harder to match.

A supplementary suggestion of introducing a new `/W`
markup has not been implemented since it would mean
added complexity in the code, the user interface and the
documentation, with no visible difference for the user,
particularly since the default blockquote CSS, and the
existing "blockquot" div CSS leave the font size unchanged,
contrary to the implication in the issue description.

Fixes #1294