Add Go Back and Go Forward to Search menu (#1344)
When GG jumps to a location and displays it roughly in
the center of the screen, e.g. through a Search or clicking
on a Checker message (not with arrow keys or similar
movements) the location is pushed on a stack. Go Back
and Go Forward navigate through that stack, in a similar
way to an Undo/Redo stack.

Fixes #916 (Buttons to be added to toolbar in #1342)