Add support for HTML curly quote conversion (#1346)
Add two new curly quote buttons, in a submenu. The first
converts straight quotes that are inside HTML tags to
obscure characters, just like ppsmq, but GG copes with
tags that are split across line breaks, such as
```
<  div class
= "test"
>
```

Second button does the reverse, converting ALL
occurrences of contour/surface integral characters back
to straight quotes.

This therefore provides at least the functionality of ppsmq,
and can probably cope better with real-life occurrences of
`<` or `>` in the text.

The reason for doing it this way rather than trying to make
both Convert to Curly Quotes and Check Curly Quotes
cope with HTML tags is that it would introduce too much
complexity into the curly quotes code, potentially breaking
it for the 99% of cases which are not HTML files.

Fixes #1216