Prune scanno/regex files tthat don't exist
If user adds their own scanno/regex files, it is likely they
may rename, move or delete them at some point. Up to this
point there has been no way for them to be removed from
the dropdown list of files.

With this commit, when the dialog is created, or when a
file is successfully loaded, the list is checked and files
that don't exist are removed.