Add Regex Library feature (#1349)
* Add Regex Library feature

1. Regex Library works like Stealth Scannos, but has its own
directory for the regex library files.
2. The file `dashes.json` is supplied with the release.

See the manual for more details:
https://www.pgdp.net/wiki/PPTools/Guiguts/Guiguts_2_Manual/Tools_Menu#Regex_Library

Fixes #663
Fixes #860

* Prune scanno/regex files tthat don't exist

If user adds their own scanno/regex files, it is likely they
may rename, move or delete them at some point. Up to this
point there has been no way for them to be removed from
the dropdown list of files.

With this commit, when the dialog is created, or when a
file is successfully loaded, the list is checked and files
that don't exist are removed.

* Fix re-run bug