Add pinning to dialogs
1. Add right-click context menu to dialogs which should
work anywhere in the dialog that doesn't already capture
a right-click.
2. Context menu shows "Pin" or "Unpin", depending on
current setting.
3. "Pinned" setting is stored per dialog in prefs file.
4. An entry in the View menu toggles the "pinned" setting
for the dialog that has focus.
5. This entry can be assigned a shortcut in the usual way.
6. When pinned, a pin shows in the title bar of the dialog.
7. When pinned, dialog is subordinate to the main window
in a way that means it remains in front, and moves up/down
the window manager stack with the main window. If the
main window is iconified/minimized, the pinned dialogs will
be too.
8. A second pinned dialog can be positioned either
above or below a first pinned dialog by clicking to raise it
as normal
9. An unpinned dialog can either be above the main window
and all pinned dialogs, or below the main window and all
pinned dialogs - it can't go "between" them
10. Pinning is not available on Linux (Window Manager
doesn't permit it)
11. The above limitations cannot be fixed - note that GG1
has the same feature via Prefs, Appearance, either for
all dialogs, or individually for the WF dialog or the S/R dialog.
GG1 has the same limitations.

Fixes #974