Add "Compress PNGs" feature
Command to run is set via Settings dialog, Advanced tab,
and defaults to `pngcrush -q -reduce $in $out`
The `$in`/`$out` indicate where the input and output filenames
are placed in the command. This is needed because some
tools, e.g. `optipng` require `out=$out` as a parameter,
and `$in` as the final argument.

If the command is empty, then the Python Pillow library
loads the image and saves it "optimized". This is likely
to be less effective than running a specialist too, but is
much quicker.

All the above is explained in the Settings dialog tooltip.

A message is output as each file is compressed, and a
summary message at the end.

The old pngs are saved in a `pngs_old` dir. If this dir already
exists, an error is popped, because that means the PPer has
already run the compression, and they may want to preserve
those original files before overwriting them when they
compress again. (guiprep has something similar)

If a file is not compressed by the tool, the original is kept,
so the results will always be at least as good as the original.