Improve ", <UPPERCASE>" regex to aid sorting
This is regex number 15 in `regex.json`. Previously only the
comma was captured and highlighted, so alpha sorting
didn't appear to work, since every match was a comma.
Changed to also capture the uppercase character after the
comma, e.g. `, C` in `, Caesar`, so sorting is based on that
character.
Note that, as previously, that uppercase `I` with a space
after it is not matched by this regex, to avoid false positives
from `Next, I followed...`

Fixes #1366