Implement custom tear-off menus (#1396)
* Implement custom tear-off menus

Settings dialog now has 3 settings (2 on Macs):
Off, Custom, Built-in (not on Macs)

As stated in the dialog, requires restart of GG so menus
can be recreated from scratch.

* Fix recent files bug (#1397)

In `master`, if run with fewer than 10 recent files in the
recent files list, then newly loaded files were not added
correctly to the recent menu until the program was
exited and restarted.

* Add guiprep functionality (#1386)

* First guiprep features

1. Import prep text files
2. Dehyphenate (doesn't use the two directory algorithm)
3. Remove Headers/Footers
4. Export prep text files

* Review improvements to dehyphenation

1. Add asterisk to end of page hyphens when dehyphenator
is run.
2. Add button to toggle Keep<-->Remove
3. Change Fix All, etc., so that "All" means all! So, if you want
to review them, you can (toggling as you go), and when
you are ready, you just click Fix All once.

* Minor display adjustment

* Update tooltip

* Further review fixes

1. Don't add asterisk if Keeping hyphen (just end of page)
2. Fix All means Fix All headers/footers with the same error
type, rather than strictly matching headers/footers.

* Add more features

1. Add "Filter" feature
2. Add "Common Scannos"
3. Add "Englifh" fixer
4. Add "[Blank Page]" fixer
5. Re-order menu options
6. Make Fix All in Dehyphenator only Fix all of matching type,
the same as Header/Footer dialog.
7. Include leading blank lines in search for header - show
them in the dialog as `⏎`

* Add "Compress PNGs" feature

Command to run is set via Settings dialog, Advanced tab,
and defaults to `pngcrush -q -reduce $in $out`
The `$in`/`$out` indicate where the input and output filenames
are placed in the command. This is needed because some
tools, e.g. `optipng` require `out=$out` as a parameter,
and `$in` as the final argument.

If the command is empty, then the Python Pillow library
loads the image and saves it "optimized". This is likely
to be less effective than running a specialist too, but is
much quicker.

All the above is explained in the Settings dialog tooltip.

A message is output as each file is compressed, and a
summary message at the end.

The old pngs are saved in a `pngs_old` dir. If this dir already
exists, an error is popped, because that means the PPer has
already run the compression, and they may want to preserve
those original files before overwriting them when they
compress again. (guiprep has something similar)

If a file is not compressed by the tool, the original is kept,
so the results will always be at least as good as the original.

* Add "Renumber PNGs" feature

It checks that the `-----File:` page separators and the PNG
files correspond. Then it renumbers the PNG files, and edits
the page separators.

When the user Exports Prep Text Files, they will be exported
with the new filenames (since they are based on the page
separator lines).

* Fix exception on empty command

* Don't insist on export prep text dir existing in dialog

It's still checked in the code.

* Re-find pngs after renumber

So that the image viewer will display them without needing
to reload.

* Fix page mark flags when renumbering

Need to delete old flags

* Allow 2-color palette PNGs to be inverted

* Also  invert 1-color paletted PNGs

* Error if png compress doesn't include $in/$out

Command must have $in and $out so code knows where to
put the filenames in the command.

* Fix $in/$out requirement

* Simplify in/out check

* Extra features

1. Add "Reset Page Marks from Page Separators"
2. Convert CP Character Subs to a Checker dialog, with
categories of substitutions, e.g. "Whitespace", "Curly Quotes",
"Hyphens", etc. Use can Fix & Hide All to substitute all of
one type.

* Move `CP Character Subs` into `Filter File`

New checkboxes added:
"Convert non-standard whitespace to space/newline"
"Convert non-standard dashes to hyphen(s)"
"Convert curly quotes to straight"

Minor change to:
"Convert, & ensure space before, ellipsis (except ....)"

Two options combined into one:
"Remove suspect punctuation from start/end of line"

Added "All On" and "All Off" buttons

* Fix review comments

1. Stop Fire being converted by Englifh, but Thanklefs should
still be converted. The rule is that uppercase characters
remain the same, but lowercase come from the corrected
English version.

2. Remove 2 additional files that slipped in during previous
testing.

* Implement custom tear-off menus

Settings dialog now has 3 settings (2 on Macs):
Off, Custom, Built-in (not on Macs)

As stated in the dialog, requires restart of GG so menus
can be recreated from scratch.

* Update application.py

Merge tearoffs & recent file fix