Fix bookmark/search cursor position clash (#1390)
If there was a selection, e.g. from a search, then going to a
bookmark did not clear that selection.

So, you didn't see the insert cursor at the bookmark
position, and if you then used an arrow key it would jump
to the location of the selection rather than moving relative
to the bookmark position.

Fixes #1374