Fix "long line" centering bug
If a line was longer than the right margin, then when it was
centered, spaces were removed from the line, including
between words (see linked issue).

Fixed by only removing leading spaces. If a line can't fit,
it is shunted to the left margin as best it can be.

Fixes #1381