Fix word boundary regex bug (#1389)
The combination of word boundary with a non-greedy
match caused the replacement to be incorrect.

Example:
S: `\b.+?\b`
R: `X`

This should find words, i.e. groups of characters delimited
by word boundaries, and replace it with a single `X`.
In fact it replaced every character in the word with `X`.

This is because we can't use Tcl regex replacement since
it is very limited, so we have to get the match into a
python string, e.g. `word`. We now execute the replacement
on that string. This means it does not have context, so
word boundaries and look-ahead/behind at the start and
end of the string don't work. We therefore removed them
from the search string used on the Python string.

However, this means that the search term is now `.+?`,
which matches just the `w` in `word`, which it replaces
with `X`, then it matches the `o`, then the `r` then the `d`.

The fix is to replace a beginning of line `\b` with `^` and an
end of line `\b` with `$`, and the same for lookahead/behind.

Fixes #1369