Add Character Suite support (#1408)
* Add Character Suite support

1. Update list of character suites and which characters they
contain from the DP website (GG1 is out of date)
2. File-->Content Providing-->Highlight WF Chars not in
Selected Suites turns on/off highlighting in the Word
Frequency dialog when in Character Count mode.
3. If a character is not in a currently selected character
suite, or not in any character suite, it is highlighted.
4. If the character suite is known, its name is printed, also
highlighted, next to the character.
5. If the character is in several suites, they are all listed.
6. If the character is in no suites, that is reported.
7. File-->Content Providing-->Manage Character Suites
allows the user to enable or disable any of the DP
character suites.
8. The list of char suites is stored in the project bin/json file.

* Turn Basic Latin on before file is loaded